adding testing

+++ std/node/_fs/_fs_rmdir.ts
@@ -1,7 +1,7 @@
 type rmdirOptions = {
-  maxRetries: number;
-  recursive: boolean;
-  retryDelay: number;
+  maxRetries?: number;
+  recursive?: boolean;
+  retryDelay?: number;
 };
 
 type rmdirCallback = (err?: Error) => void;
@@ -10,19 +10,21 @@ export function rmdir(path: string | URL, callback: rmdirCallback): void;
 export function rmdir(
   path: string | URL,
   options: rmdirOptions,
-  callback: rmdirCallback
+  callback: rmdirCallback,
 ): void;
 export function rmdir(
   path: string | URL,
   optionsOrCallback: rmdirOptions | rmdirCallback,
-  maybeCallback?: rmdirCallback
+  maybeCallback?: rmdirCallback,
 ) {
-  const callback =
-    typeof optionsOrCallback === "function"
-      ? optionsOrCallback
-      : maybeCallback || (() => {});
-  const options =
-    typeof optionsOrCallback === "object" ? optionsOrCallback : undefined;
+  const callback = typeof optionsOrCallback === "function"
+    ? optionsOrCallback
+    : maybeCallback;
+  const options = typeof optionsOrCallback === "object"
+    ? optionsOrCallback
+    : undefined;
+
+  if (!callback) throw new Error("No callback function supplied");
 
   Deno.remove(path, { recursive: options?.recursive })
     .then((_) => callback())
