+++ std/node/_fs/_fs_unlink.ts
@@ -1,4 +1,5 @@
 export function unlink(path: string | URL, callback: (err?: Error) => void) {
+  if (!callback) throw new Error("No callback function supplied");
   Deno.remove(path)
     .then((_) => callback())
     .catch(callback);
